Fix equality check for sparse, move asarray to utils
Previously, checking equality during concatenation between a sparse array and a dense array or dataframe would throw an error.
This was causing intermittent test failures from `test_concatenate_uns`.

Now equality of values is checked for. It's arguable that arrays of different types should compare false, so maybe this should be changed.

I've also moved `asarray` from `tests.helpers` to `utils`, since it is a generally useful function that isn't specific to the tests.
