+++ anndata/tests/test_obspvarp.py
@@ -6,7 +6,8 @@ import pytest
 from scipy import sparse
 
 import anndata
-from anndata.tests.helpers import asarray, gen_typed_df_t2_size
+from anndata.tests.helpers import gen_typed_df_t2_size
+from anndata.utils import asarray
 
 M, N = (200, 100)
