+++ anndata/tests/test_readwrite.py
@@ -10,8 +10,9 @@ import pytest
 from scipy.sparse import csr_matrix, csc_matrix
 
 import anndata as ad
+from anndata.utils import asarray
 
-from anndata.tests.helpers import gen_adata, asarray, assert_equal
+from anndata.tests.helpers import gen_adata, assert_equal
 
 HERE = Path(__file__).parent
