+++ anndata/tests/test_views.py
@@ -8,13 +8,13 @@ import pytest
 
 import anndata as ad
 from anndata._core.index import _normalize_index
+from anndata.utils import asarray
 
 from anndata.tests.helpers import (
     gen_adata,
     subset_func,
     slice_subset,
     single_subset,
-    asarray,
     assert_equal,
 )
